Remove Broadlink, Add AV Receiver, Guest Mode Tweaks, Vacation Mode Tweaks
My Broadlink RM Mini 3 died to due toddler abuse.  I found a great deal on a AV receiver that is network connected so it replaced most of what the RM Mini 3 was used for.  I also went through and added conditions for guest mode and vacation mode (input booleans for each) rather than running an automation to turn on/off applicable automations.